Replace assert log method with context manager
